Move resource libraries to the main libraries directory otherwise chef would fail to compile the cookbook
