+++ chef/cookbooks/pacemaker/libraries/resource_ms.rb
@@ -1,0 +1,11 @@
+
+module Pacemaker
+  class Resource::MasterSlave < Pacemaker::Resource::Clone
+    TYPE = "ms"
+    register_type TYPE
+
+    #include Pacemaker::Mixins::Resource::Meta
+
+    attr_accessor :rsc
+  end
+end
